+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 ## [next]
 
 * Correctly paint buttons borders ([#956](https://github.com/bdlukaa/fluent_ui/pull/956))
+* **MINOR BREAKING** Removed `ButtonStyle.border`. Use `ButtonStyle.shape` instead:
+  Before:
+  ```dart
+  Button(
+    style: ButtonStyle(
+      border: ButtonState.all(BorderSide(...)),
+    ),
+  ),
+  ```
+
+  Now:
+  ```dart
+  Button(
+    style: ButtonStyle(
+      shape: ButtonState.all(RoundedRectangleBorder(...)),
+    ),
+  ),
+  
+  ```
 
 ## 4.7.7